Tests & WIP redirecting / to relevant team pages
I haven't been able to try this at all because my checkdesk
app is currently broken, so it might be wildly incorrect, but
hopefully the intent is clear from the code. 

In case not: We want to send users from / to /teams/new if they
don't have a team yet. If they do, send them to /team/:current.