Fix client test for adding duplicate tags
This test seemed to assume that the media already had a tag
from previous tests (something we want to move away from),
and then would test if re-adding the tag would be prevented.

Instead, in concert with Rspec Retry, it seems to always
fail the first time (because the assumption is untrue today),
as well as adding a tag in the process, which then makes it
succeeding the following time. (Example: any recent test output)

This change ensures assumptions are valid before it proceeds.