Node unit tests: fix; load test config.js more smoothly
Also, upgrade mocha and sinon.

Previously, config.js was loaded by copying it into `node_modules/` -- which
is a no-no because it pollutes the environment. Now, we override require()
to load the test config.

When we upgrade to Jest, we can use jest.mock(). For now, this is pretty
darned close.