Move `pusher` to a React.Context.
Initialize a single instance in `app/pusher.js`. This is kinda like
a global, except we can mock it in unit tests by passing
`props.pusher`

This doesn't accomplish much cleanup on its own. But it's a start.
Next, all these Pusher subscribe/unsubscribes should happen in
higher-order components so we can get rid of CheckContext.