<MoreLess>: pre-emptive bugfixes
The previous logic compared `contentHeight` with `areaHeight`. Problem
is, `areaHeight` depends on `contentHeight`! So I moved the hard-coded
`max-height` into separate variable, so we can test `contentHeight`
against the hard-coded max-height.

Also, added `key` to each caller. That way, when a new batch of content
is rendered, the state will be collapsed. (Imagine flipping through
media pages with the "next" button and having the component tree stay
intact. We'll get there one day.)