<Home>: nix `state.open`, which is unused
<TeamHeader>: nix `onClick`, which modified <Home> state.open. Rewrite
without styled-components. (The code makes it look as though <Offset>
moves content to the right; but it doesn't. So don't use <Offset>.)

I had an extremely hard time actually _rendering_ <TeamHeader>. I
removed a redirect from `CheckNetworkLayer.js` and browsed to a media
with a valid ID (because an _invalid_ ID would lead to redirect).