Revamp <TeamMembers>. Eradiate <Offset> and isRtl.
Update `react-icons` (for Slack) and use new `<SocialIcon>` throughout.
Fixes a failing prop-type. [#8179]

The isRtl removal is for #8347 -- nixing the "intl" variable everywhere.

Then <Offset> ... I don't think anybody understood what it did. It
was often added as the last item of an align-left container, and all it
did was add a padding-right. That's zero value. It depended on `isRtl`,
so it had to go. [#8347] [#8249]

To nix <Offset> from <TeamMembers>, I used material-ui's list-item
components [#8249] ... and I noticed that the <RCTooltip> was putting
tooltips in the wrong place because (A) they were exactly where the
mouse goes when switching between users; and (B) they would _load_ and
then switch size after loading. I re-examined the Relay code there. This
should make it easy to switch to Relay Modern [#6211].