Feature/8347 react intl prepping for upgrade (#646)
* [#8179] CreateMediaInput.test: s/render/mount

* CircularProgress: overflow:hidden

* [#8179] Login.test: mount(), not render()

* Nix skipped test

* QuoteMediaCard: fix unit-test warnings [#8179]

* Change away from classNames to HTML-standard "dir".
* Add "lang", while we're at it, for HTML compliance.
* <MoreLess>: use refs, not `document.querySelector`, so it works in
  unit tests.
* <MoreLess>: change state to store raw values (`areaHeight`,
  `contentHeight` instead of derived values (`canExpand`).
* <MoreLess>: don't listen for "resize" events on non-`window`, because
  they'll never fire.
* <MoreLess>: `s/shouldComponentUpdate()/React.PureComponent`

* ParsedText: fix unit-test warnings [#8179]

* Sentence.test.js: fix warnings [#8179]

* <MoreLess>: pre-emptive bugfixes

The previous logic compared `contentHeight` with `areaHeight`. Problem
is, `areaHeight` depends on `contentHeight`! So I moved the hard-coded
`max-height` into separate variable, so we can test `contentHeight`
against the hard-coded max-height.

Also, added `key` to each caller. That way, when a new batch of content
is rendered, the state will be collapsed. (Imagine flipping through
media pages with the "next" button and having the component tree stay
intact. We'll get there one day.)

* <PageTitle>: nix skipTeam

* Nix Button.hoverColor prop [#8179]

It had no effect and nobody has noticed a problem.

* <UploadImage>: remove state [#8362]

Callers all responded to `<UploadImage>` value changes as though it was
a controlled component; but it was _implemented_ like an uncontrolled
component. This commit fixes all callers and nixes lots of use of
`document.forms`.

* Typo fix

* Fix race in integration tests

* rule.spec.js: eslint

* Fiddle with some integration tests

* Login.js: try and fix integration tests

No more "toggle" buttons or "toggle state" test instructions: now
there's a "register" button and a "login" button.

* Nix appium_lib [#8228]

This removes tons of noisy warnings from Travis output.

* Another #register in integration tests

* Nix all references to BrowserStack

It's unused. And if we wanted to use it today, we'd surely need to
change things a lot.

* Specs: use local variables, not global ones

* Nix appium from Gemfile.lock

Fixes warnings on integration-test startup [#8179]

Also, update Selenium in Gemfile.lock.

Nixing appium removes `Selenium::WebDriver::Element#value`. Appium adds
the method, and Selenium docs describe it but don't actually use it.

* Wrap LoginContainer with <div>, not <p> [#8179]

* ReportDesigner: avoid derived state

Fixes warning [#8179].

What started as an update for <UploadImage> changed into so much
more.

`<ReportImagePreview>`: avoid HTML injection. Use a single `image`
prop. Switch to Hooks.

* Integration tests: simplify after Appium removal

* `new_driver` doesn't need arguments
* Use `--disable-dev-shm-usage` so Chrome won't crash in Docker
* `wait_for_selector()` et al: cleanup. Don't rescue all errors.
  Use the same logic throughout all functions. Don't double-loop.
* `wait_for_size_change()`: fail if size does not change.

* Revert "Integration tests: simplify after Appium removal"

This reverts commit 338dbb90c81670605a61d27430d7c65b844a4304.

* More `.value` replacements in specs

* Make a test less flaky

* Nix another injectIntl() [#8347]

* ReportDesigner: correctly parse create response

* SearchQueryComponent: cleanups

* Nix swallowingStyled [#8249]
* Nix injectIntl [#8347]

* eslint: go all the way with <LocalizedRole>

* SwitchTeamsComponent: Follow material-ui docs

* `<ListItem>` should be a direct child of `<List>` [#8249]
* `<ListItem>` `component` prop is `Link`
* Use `<ListItemAvatar>` to display avatar with correct spacing
* Use `<ListItemSecondaryAction>` `<IconButton>` to display action icon
  correctly.
* Nix `isRtl` [#8347]

* Wire up material-ui RTL. Nix *lots* of isRtl

[#8347] [#8249]

Lessons of the day:

* We overwrote a couple of material-ui defaults here. For instance,
  <Drawer> will appear on the correct side by default, and rtlDetect was
  giving us the wrong logic. (Also, we weren't setting the drawer width
  correctly -- fixed!)
* There was lots of copy/pasted code.
* `text-align: start` and `text-align: end` align buttons perfectly
* `flex-grow: 1` aligns block contents perfectly

* <Home>: handle <FormattedMessage> messages

* <Home>: nix `state.open`, which is unused

<TeamHeader>: nix `onClick`, which modified <Home> state.open. Rewrite
without styled-components. (The code makes it look as though <Offset>
moves content to the right; but it doesn't. So don't use <Offset>.)

I had an extremely hard time actually _rendering_ <TeamHeader>. I
removed a redirect from `CheckNetworkLayer.js` and browsed to a media
with a valid ID (because an _invalid_ ID would lead to redirect).

* <Home>: add classes for integration tests

* Fix import

* <TeamInviteCard>: don't use RTL [#8347]

Also, use material-ui CSS [#8249]

* Revamp <TeamMembers>. Eradiate <Offset> and isRtl.

Update `react-icons` (for Slack) and use new `<SocialIcon>` throughout.
Fixes a failing prop-type. [#8179]

The isRtl removal is for #8347 -- nixing the "intl" variable everywhere.

Then <Offset> ... I don't think anybody understood what it did. It
was often added as the last item of an align-left container, and all it
did was add a padding-right. That's zero value. It depended on `isRtl`,
so it had to go. [#8347] [#8249]

To nix <Offset> from <TeamMembers>, I used material-ui's list-item
components [#8249] ... and I noticed that the <RCTooltip> was putting
tooltips in the wrong place because (A) they were exactly where the
mouse goes when switching between users; and (B) they would _load_ and
then switch size after loading. I re-examined the Relay code there. This
should make it easy to switch to Relay Modern [#6211].

* Nix unneeded `injectIntl()`s

* <AutoCompleteMediaItem>: nix `intl` and fix races

Previously, two autocomplete HTTP requests could race one another. It
looks like the component used a 2-second search timeout to avoid most
races -- hurting usability and not even fixing the problem.

Now, changing search text aborts the previous autocomplete.

Also, the "Loading" and "Error" messages are built-in to the component
now: that's part of the material-ui <AutoComplete> API.

* Fix unit test

* <MoveDialog>: intl [#8347], Relay [#6221]

* Disable unused, untested API for choosing from a different team.
* <DestinationProjects>: use <Autocomplete> instead of react-select.
  This fixes a warning [#8179] and cleans styles [#8249].
* Make everything a fragment container [#6211]. Reduces code spaghetti;
  nixes a dependence on `CheckContext`; saves an HTTP request.

* <MediaActionsBar>: nix `props.intl`

* <MediaCondensed>: nix props.intl

* <MediaStatusCommon>: nix props.intl

* <BulkActions>: fix query fragment

* <DestinationProjects>: auto-highlight, auto-focus

Now it's even better than react-select ever was.

* <MediaTags>: nix props.intl

* smoke_spec: fix <Autocomplete> interaction

* Nix unused import

* Nix file ... how did it get there?

* Nix now-unneeded `direction` from CSS

* <MediaActionsBar> Fix spaces in titles/descriptions

* Un-delete integration-test line

* Fix a couple of tests

* Nix cruft (thanks @caiosba)

* Integration-test fix

* <SearchResults>: add a forceFetch

It should startle us that most integration tests passed without this
line. I suspect that's because the integration tests all interfere with
one another.