Conditional fields in Annotation Settings
This change fixes an issue where annotation settings were being updated to an old state, by triggering the save on a `useEffect` whenever any of the conditonal fields have been updated by the user. This uses a trick with a `firstUpdate` useRef boolean in order to not have the save trigger on initial page load (otherwise we'd get a superfluous POST to the API).