CHECK-937: conditional annotation "save all" and "required" fields (#922)
* CHECK-937: Annotation: save all & required fields

Overall this is a clientside implementation of "save all" which gets around the fact that we don't have that functionality on the graphql DB. We now hide the save/edit/delete buttons on the metadata fields, but they still exist in the DOM. We iterate over each hidden save button and check html data-elements in order to see whether a field is empty and whether it is required. If a field is required and empty, we don't allow the saves to happen. Otherwise, we simulate a click on each save button, thus saving all the fields.

- add a function that evaluates all metadata fields in the form and finds the most recently updated time and user who updated it
- use html data-elements to indicate on each Save button whether the fields are required or empty
- store a local copy of the overall metadata fields and mutate that for local display, applying those changes to the DB when we save (conditional display of fields uses this local copy)

* Updating integration tests for save-all annotation

* Updating custom folder test for new annotation entry