CHECK-1035: filters not applying in tandem (#936)
* Convert SearchResults to functional component

This is preparatory work for CHECK-1035, modernizing and reducing the complexity of the `SearchResults` component in anticipation of changing how state is managed on its children.

* Completing PropType definitions

* SearchKeywords is now stateless for query

It relies on a query state passed down from its parent as a prop, which allows it to share state with other components like `SearchFields`.

* SearchFields component query is now stateless

`SearchFields` is now passed the `query` value as a prop, which is a state managed on its parent `SearchResults`. The state is shared between `SearchFields` and `SearchKeyword`.