CHECK-1088,1091 Fix attribution of annotations
We were looking at the wrong fields in the graphql DB to get the annotation times. We now grab `annotator` under `first_response` on an individual task instead of `annotator` on the root projectmedia.

We also no longer update an annotation when its current local value is the same as the value on the server (aka it hasn't been edited).

Fixed a minor bug with missing `about.file_extensions`.