Use setRouteLeaveHook to prevent in-app nav
During editing, we now prompt on the use of the browser back button and the in-app nav button. Since the prev/next item buttons are now disabled during editing, the current issue with not knowing the next URI before making an API call is effectively avoided. Eventually we will be able to pre-compute the URI for the prev/next buttons and use this exact setRouteLeave solution to prompt users on navigation (since we can diff the routes) but that is for CHECK-1117.