Rewriting due to test failures
Ripped out the context stuff, went back to basic useState, ended up doing a bit of DOM inspection when the user clicks on the prev/next item indicators to prevent the loading state from happening if we are currently editing annotations. This prevents rerendering of the entire MediaPageLayout component and prevents a bug where tabs would switch back to Annotations from Source if you tried to edit a Source Annotation.