Navigation prompt fix (#977)
* Use setRouteLeaveHook to prevent in-app nav

During editing, we now prompt on the use of the browser back button and the in-app nav button. Since the prev/next item buttons are now disabled during editing, the current issue with not knowing the next URI before making an API call is effectively avoided. Eventually we will be able to pre-compute the URI for the prev/next buttons and use this exact setRouteLeave solution to prompt users on navigation (since we can diff the routes) but that is for CHECK-1117.

* DOM inspection hack due to NextPreviousLinks architecture
Ended up doing a bit of DOM inspection when the user clicks on the prev/next item indicators to prevent the loading state from happening if we are currently editing annotations. This prevents rerendering of the entire MediaPageLayout component and prevents a bug where tabs would switch back to Annotations from Source if you tried to edit a Source Annotation.