CHECK-1115
This fixes two problems. One, there were cases where empty multiple choice or file upload fields were being submitted to relay as a mutation of a non-existent existing record. This was due to an error in the logic flow for determining whether something is a "create" or an "update" or an "ignore". That logic has been fixed. Two, this provides a backup for legacy data where an item was updated without an original item existing -- we were missing an `id` field needed for the graphql update to work, so now if that field doesn't exist where we expect it, we look for it in the place where it must be, in the legacy data.