CHECK-1266 - fix missing navigation buttons on similar media (#1008)
When we link to the similar media sub-page, we don't pass along the `listIndex` url parameter and its value, nor do we do so on navigating back to the parent page. This causes the next/prev navigation field at the top to fail to display when you go to the item, and when you go back. This commit rectifies that error.