Fix CHECK-1390
The `TableContainer` was flowing off the bottom of the screen, which resulted in both the horizontal and vertical scroll bars being wrong (but only horizontal prevented content rendering, so this bug only caught that case). By bringing the "bottom" of the `TableContainer` "up" more (increasing the negative vertical offset of `max-height`), we constrain the table to the viewport and get the correct scroll behavior overall.