CHECK-1418 Add truncation to source in column view
When "Source" is shown in search results, we now truncate to 25 characters and put `'...'` after the text.

This commit also removes our dependency on `lodash.truncate` in `helpers.js`, replacing it with a couple lines of plain JS. I've also found locations where we use a custom truncate in the code and updated so we use the helper function instead.