Refactor/1642 browser back button (#1096)
* Fix browser back button double-click issue

We had a bug where clicking a `SearchResultsTableRow` item and pressing the browser back button didn't work -- you had to press it twice for it to work. Pressing the in-app back button only required one press, though. It took a while but I figured out the reason.

We don't need to push to `browserHistory` on a click of a `SearchResultsTableRow` -- this is a little weird, but the `TitleCell` is rendered wrapped in an href (a `Link` component) which you would expect to only work when you click the `TitleCell` and not the other columns. But! `TitleCell` is rendered as `<th>` instead of `<td>`, which applies the link to the entire row! It's the only cell in the search results table row that is `th` and overrides everything else.

This means that if we pushed to browser history on a click event for the row, it would first apply the href to the history, then push the history (or maybe in the other order). The result is two items on the history stack for one click. Clicking the in-app back button simply pushes the correct url so it doesn't matter how much stuff we have in the stack, but clicking the browser back button *pops* the history, which means we have to pop twice, which is two browser back button clicks.
 - This also addresses a potential source of future bugs by getting rid of a "magic number" in the code. We don't want to query 50 items at a time in `MediaSearchRedirect`. We want to query `pageSize` items at a time.