Keep track of number of in-progress mutation requests
On the network layer, we modify a global `inFlightMutationRequests` variable that lives on `window` and can be accessed from any function. This was the most reasonable solution I could come up with for a small amount of code, even though it doesn't hook into the React lifecycle. The `HomeComponent` now calls `setRouteLeaveHook` with a function that tests to see if the number of in-progress mutation requests is greater than 0, and prompts the user to see if they really want to navigate away.

In the Jira ticket, the wording was to do this whenever _queries_ are still processing, but it doesn't make sense for non-mutation queries. A user should be allowed to navigate away from a page mid-READ, but prompted mid-WRITE.

This also removes an extraneous `withRouter` no longer needed in `Tasks.js` after Alex refactored it to use the `NavigateAwayDialog`. I am not using that new component here because this is a special case where I want the `setRouteLeaveHook` logic to be different and not coupled to the state change of a particular component. If I did that to `HomeComponent`, we would end up constantly rerendering the entire application, so it doesn't make sense here.

Fixes CHECK-1643.