Always show "Submitted" column setting (#1110)
We made a design decision to always show the "Submitted" column setting. But there are some cases where workspaces that are older than our decision had it explicitly turned off, which means its `show` value was `false`. This commit makes a column setting show if it has the correct key value for the "Submitted" column, `created_at_timestamp` (the key is independent of localization).

I added a unit test for the functionality where the "Submitted" column setting should lack a show/hide button.

We are also replacing several integration tests with unit tests here. There is an integration test that checks to see if clicking the "Submitted" field changes the sort direction. The actual sorting is up to `check-api` and not `check-web`, so assuming that sorting is well-tested on `check-api`, this is now replaced with a unit test that makes sure the UI widget displays the correct direction when sort direction is changed via the URL (since ultimately all sorting action triggers a url change, then a call to the API, then a render of elements clientside). I've also removed a test (`should search and change sort order`) that was redundant to the unit test once we consider how the search and sort logic works.

Writing the unit tests also caught a small bug with unsafe object property access that has been fixed.