Update default sort on suggested matches page
This commit changes the default sort on the `/suggested-matches/` route to "Submitted" (`recent_added`).

Added a unit test and picked up and fixed an additional minor bug in the `propTypes` for the `Search` component.

Note on testing: by running the `render` component of a `ReactRelayQueryRenderer` and passing in the appropriate `error` and `props` in a object as an argument to `render`, we can mock the relay renderer and do things like inspect derived properties on the rendered component.

Fixes CHECK-1782.