Improve search input performance
Altered `SearchField` so it only updates the parent `query` value onBlur (on on an 'Enter' keypress) instead of onClick. This avoids rerendering the entire search page on each button press.

This also removes a side effect that caused the `<title>` of the page to update with each keypress (since the `PageTitle` is bound to `query` in `SearchKeyword`).

Add unit tests for basic page title rendering.