Improve search input performance (#1117)
Altered `SearchField` so it only updates the parent `query` value onBlur (on on an 'Enter' keypress) instead of onClick. This avoids rerendering the entire search page on each button press.

This also removes a side effect that caused the `<title>` of the page to update with each keypress (since the `PageTitle` is bound to `query` in `SearchKeyword`).

Add unit tests for basic page title rendering. We export the core `SearchKeyword` component (as `SearchKeywordTest`) without its graphql fragment container, allowing us to mock any of the variables that would be passed by graphql. This way we are testing the component itself and not Relay.