Fix Smooch bot resource editor regression
Fixes this by getting rid of randomized key props for TextFields in SmoochBotResourceEditor.js and adding a key to SmoochBotResourceEditor based on the resource id from the parent component.
This should eliminate the need for the useEffect call to reset state when browsing through resources, so removing this too.

Fixes CHECK-942