Show placeholder tag when deleted filter selected (#1125)
If a user did something like create a new Status type, make a filter based on that Status, and then delete the Status, the filter would still show up just as a blank. This replaces the blank with a red "Property deleted" tag to notify the user.

Added tests, and added a `checkError` color value since we have a new "brandError" red color in Figma. Fixed some missing PropType declarations, too.

Fixes CHECK-1396.