Provide default sort to SearchResults
While we do set a default sort at the UI level in `SearchResults`, we still need to set the default sort on the higher-order parent component `Search` so that the sorting logic is applied to the data query itself. Otherwise we end up passing an unsorted set of data into a table that looks like it has sorted the data, but it hasn't.

Fixes CHECK-1863.