Hotfix to sorting issue
When smoochbot is enabled, `check_api` returns items by default sorted by `last_seen`. This information was missing from the default graphql query, so we add the item and then make sure that the `SubmittedCell`, which is supposed to show the more recent of `last_seen` and `created_at_timestamp`, displays the correct value. This way the sort is correct and we are also not invalidating the relay cache by sending it a manual sort value.