On report designer, show either the text report or the image report
This is mostly for legacy reasons, since today it's not possible, on the frontend side, to create both text report and visual card for the same item. But for some existing data, some items have both a text report and a visual card. Let's be sure that only one is displayed for those cases.

Fixes CHECK-1897.