Limited text fields with counters
Adding a new component that is a text field that has a maximum length and that displays a counter in real time of how many characters remaining until the limit is reached. Refactored fact-check and report to use this new field. Includes a unit test for the new component.

Reference: CHECK-1925.