Mandatory description prop for FormattedMessage components.
Now in addition to id and defaultMessage please add a description field to your FormattedMessage components with context for the translators. Translators often need to know how to agree to gender and number. It is helpful to state what a string refers to, especially when it’s a single adjective. E.g.: Say the string is “Similar”, then add to description: “Label that refers to items marked as similar to the current one”.

Fixes CHECK-1402