Request team task "tasks_with_answers_count" field only on demand.
The GraphQL field "tasks_with_answers_count" from "TeamTaskType" is pretty expensive. It was required in the list view of team tasks (a.k.a., "Annotations" settings page for a workspace) for all annotation fields, although it is required only when the "edit" or "delete" options for a field is clicked. So, there was a N + 1 performance problem happening here for workspace with many annotation fields. The fix was to wrap the components that need the "tasks_with_answers_count" GraphQL field by a new container that requests that field.

Reference: CHECK-2089.