Add disabled relay/unused-fields rule to files
Per discussion on the engineering team, I'm adding the eslint relay plugin and having it flag an error whenever graphql fields are not used _in the same file_ as the query/fragment.

This means that the dozens of places where we have a big query on a parent component and then pass the result down to a child in a different file _will_ trigger a linter error. Since we can't fix this all at once, I'm taking the same path we did with the `description` eslint rule for react-intl and selectively disabling it in all files. As we do work on these files in the course of normal Check development, we will refactor the queries to correctly follow a pattern where data fragments of minimal size are pulled in only when needed by the code.

I also took the time to remove a few unused fields in a few files. I also added some `description` fields in a few places and removed the corresponding `eslint-disable @calm/react-intl/missing-attribute` comment in the header.