Feeds v1
This is an evolution/refactoring of the previous "shared database" concept, which had hard-coded rules. In order for a search to return data from multiple workspaces, those workspaces had to have a `trends_enabled = true` setting and should have the same value for their `country` column. All the logic related to "country" and "trends" was deleted in this commit, because it was used only for this. 

For the refactoring part explained above, every instance of "trends" was renamed by "feed".

I added some new components. The most important one is `<Feed />`. It is accessible by a route `/:team/feed/:id/:tab`. It contains two tabs: "shared" and "feed". That `<Feed />` component is basically a wrapper for the `<Search />` component passing different `props` to it. There are some possible situations:

* If the user has no access to the feed or it doesn't exist, redirect to a "not found" page
* If the the tab is "shared", show content from that workspace that is shared in the feed
* If the tab is "feed" and the feed is `published`, show fact-checks from the feed (the columns for this view are hard-coded)
* If the tab is "feed" and the feed is not `published`, show clusters from the feed (WIP)

The "feed" tab is not displayed if "Enable sharing" is off. It can be turned on from that interface as well. Filters can be saved too. In that interface, "AND" is a mandatory operator. Since some filters are defined by the feed and can't be changed at a workspace level, some filters are marked as read-only. In order to do that, a `readOnlyFields` prop was added to `SearchFields`, and the values are propagated to each field, in their `readOnly` prop.

Feeds are listed in the left sidebar, in a collapsible panel, just like folders and lists. The panel is not even displayed if the workspace has no feeds. By default it opens the feed in the "shared" tab.

I implemented unit tests for the new components and I also added missing "description" attributes for the `<FormattedMessage />` components in the components I worked on.

Reference: CHECK-2102.