Refactor of MediaSimilaritiesComponent
This is an atomic commit meant to demonstrate a complex refactor led by the `relay/unused-fields` eslint rule. This commit takes `MediaItem` and rips out all of its interactive features (a pin/detach menu) and makes it into strictly a display component meant to show a summary of a media item in a card.

The interactive features of `MediaItem`, which have to do with pinning and detaching re
lationships in similar items, are moved into a new sibling component called `MediaRelationship` and a subcomponent called `RelationshipMenu`. Specifically, this refactor better utilizes fragments and avoids the pattern where we query for very large objects on a top level component and pass them down the component tree. We now pass individual properties of these objects where it makes sense, and otherwise we query for those objects on an as-needed basis as graphql fragments.