Refactor MediaTags relay/unused-fields
This commit is another good example of using the `relay/unused-fields` rule to refactor a component. This was driven by removing the eslint directive from the header of `MediaTags` and observing that the `tags.node.tag` and `suggested_main_item.dbid` fields are never used. For the latter it was a quick stylistic fix to check for truthiness via the presence of an object with a `dbid` (rather than the presence of an empty object, which is conceptually problematic).

The other error regarding `tags.node.tag` led to an investigation of the contents of the child component `TagMenu`. This component uses `tags.node.tag` but it actually does so in its own fully-hydrated query! It turns out we only really need to pass `projectMedia.dbid` down to `TagMenu`, and then `TagMenu` will handle the rest in its own query. We also discovered that `TagMenu` was checking for a `project_id` but that this is vestigial code and no longer needed, so that check was removed as well.