Change accept/reject flow for child suggestions
This commit adds a new `MediaSuggestionReview` component that conditionally renders in `MediaSidebar` when the user is viewing an item that is either confirmed similar to another item or suggested as similar to another item. It replaces the middle panel accept/reject widgets and allows the user to make a quick comparison to the `MediaClaim`, `MediaContext`, and `MediaFactCheck` in the left pane _or_ to open the suggested similar item in a new tab for more detailed comparison. When a user accepts similarity, they are prompted to move to the parent item. When a user rejects similarity, the suggestion is deleted and the user can simply work on the item as normal.

Building this as a `createFragmentContainer` in our new Relay style would have been too arduous a task as the `Media` component itself needs a ton of refactoring. This will be its own ticket to come in Jira.

CHECK-1687