Updating integration test for custom columns
This used to be dependent on the number of default columns in the settings page. Now instead of counting to the Nth element, where N can change based on other features, we use an xpath that looks for the word "metadata", jumps up 3 levels, and then goes down 2 levels to a sibling button element that it can click.