Prevent random argument reset
This commit is a workaround for the random argument being reset every time the filters are interacted with.
Keeping the random value in a state generates it only once when the component is mounted, as opposed to regenerating on each render
which would make the QueryRenderer refetch unwantedly.

References CHECK-2366 CHECK-2331