Matched media scrolling and CDS implementation
This commit changes the behavior of the "Matched media" button in the similarity panel to scroll down to the media list instead of loading a new page. This commit also attempts a strict implementation of a Check Design System (CDS) component for the first time -- there is a new directory at `components/cds` that is organized to match the Figma CDS document organization.

Components in the CDS directory are meant to be strictly view components, with business logic and data passed into them as props. FragmentContainer components will still live in the usual places in the code but should import CDS components to render data retrieved from the API wherever possible.