Matched media scrolling and CDS implementation
This commit changes the behavior of the "Matched media" button in the similarity panel to scroll down to the media list instead of loading a new page. This commit also attempts a strict implementation of a Check Design System (CDS) component for the first time -- there is a new directory at `components/cds` that is organized to match the Figma CDS document organization.

Components in the CDS directory are meant to be strictly view components, with business logic and data passed into them as props. FragmentContainer components will still live in the usual places in the code but should import CDS components to render data retrieved from the API wherever possible.

Also removing several `eslint-disable` localization directives that are no longer needed.

And updating integration tests because there is no longer a separate "Matched media" page. The tests can simply check for the presence of a certain number of `.media__relationship` items (matched media items, 1 less than the "N medias" text in the previous tests because the "N medias" counted the original media item).

CHECK-2057, #1263