WIP - matched media modal
This commit is implementing the "Media and requests" modal seen here: https://www.figma.com/file/rnSPSHDgFncxjXsZQuEVKd/Check-Design-System?node-id=495%3A8955

This implements that modal as a CDS components as `MediaAndRequestsDialogComponent`, and also implements the MainButton CDS component.

This also creates an animation that highlights the matched media items when the "Matched Media" CounterButton is clicked.

One major TODO is to change how selecting media works so that the modal is instantiated from inside MediaRelationship rather than the weird state variable on MediaSimilaritiesComponent! This would mean that we could simply pass down the functionality needed for the various buttons in the modal from what would now be its parent component, MediaRelationship.