Refactor how MediaRelationship selection works
Instead of storing a state variable on `MediaSimilaritiesComponent` containing the id of the selected `MediaRelationship`, we now simply track an `isSelected` state inside `MediaRelationship` that is trigged via `onClick`. This allows us to implement "Pin as match" and "Un-match" buttons on the `MediaAndRequestsDialogComponent` using the same code as the same functions in the `RelationshipMenu`.