Use Request card in tipline requests tab
This commit refactors the TiplineRequest component to use the Request component made for the cds - Check Design System.
Also RequestCard.js is renamed to Request.js and moved to the cds folder following the organization structure on Figma design.
Minor adjustment on BulletSeparator component to match the design font size.

References CHECK-2495