Hotfix for tabs not switching in certain paths
When the user is on a `similar-media` URL path, the state of `showTab` in `MediaComponent` was always forced to be `suggestedMedia`. This fixes the state so it simply sets `suggestedMedia` to be the initial state of the tab.

NOTE, TOFIX: There is a big relay bug where the wrong query is being rendered for a component due to some issues that I'm not sure about. After talking with Caio I am releasing this hotfix that makes our graphql slightly more confusing but fixes the bug. We will later figure out the actual bug and remove this hack.