Fixes to small media cards
Making some CSS fixes. Also we were missing a potential piece of data for Link media cards, which was causing the link to not render in the `MediaCardCondensed` component, so I changed the logic of what it's looking for in its `media` property and also changed the graphql query that creates the `media` item we pass to it for this feature.

Also updated `BulletSeparator` to have a `compact` component -- eventually this will be ported to Check Design System and not necessary anymore but for now this solves the layout here without affecting layout elsewhere in Check.

CHECK-2694