List languages from data in data dashboard
Today, the data dashboard displays data by language. The language options in the drop-down are the workspace languages. So, if some language is deleted from the workspace settings, then the data for that language won’t be displayed anymore. We can avoid that by listing the language options from the languages we have data for.

Fixes CV2-2698.