Paginated, bulk UI for accept/reject/spam/trash suggested media (#1325)
First of all, this PR contains a refactor of `MediaSuggestionsComponent` such that it is now paginating the data for suggested similar media. The component is now the child of `PaginatedMediaSuggestions`, which is the child of `MediaSuggestions`. Both of the two latter components are graphql wrappers and all application logic is in `MediaSuggestionsComponent`.

The overall reasoning and methodology is explained in detail on our Confluence wiki in the "How to paginate in Relay 1.7" article in the Engineering Cookbook. That page also explains some Relay 1.7 compiler quirks that required me to duplicate some queries instead of referencing them. The self-contained commit for pagination is a17a004.

I wrote some unit tests for `MediaSuggestionsComponent` around pagination. 

There is a `SuggestedMediaDialogComponent` which is the popup you get when you focus on a single suggested item in the suggestions list by clicking on its card. In the full design, this is its own paginated component that lets the user navigate item-by-item through the entire list -- it should be its own ticket, and I'll write that one up.

### One big file (future refactors)

In general, `MediaSuggestionsComponent` is way too long, at 1069 lines of code. There are a few obvious places to address this.

There are 6 dialog boxes, separately defined, that comprise about 100 lines of code. Componentizing these dialogs is necessary but we really should do them such that they are CDS components -- these would be the "Custom vertical" and "Import data to workspace" components under "Menus, lists and Dialog" in Figma. Ultimately this needs to be turned into a CDS dialog box component and it should be six lines of code here like

```
<ConfirmationDialog title={foo} confirmButtonText={bar} onConfirm={doThing} />
<ConfirmationDialog title={baz} confirmButtonText={biz} onConfirm={doThing2} />
<ConfirmationDialog folderList={folders} title={title} confirmButtonText={text}  onConfirm={doThingWithFolders} />
```

...or something like that. Per CDS we have a `Dialog` component defined with six variants, but maybe there is some work to be done around conceptual grouping.. for example some of these are pure display components and some of these have forms embedded.

Anyway, instead of componentizing this early, I figured I would copy/paste code here and make a bookmark to come back when it's time to work on developing the confirmation dialog component.

Another place we should refactor `MediaSuggestionsComponent` is moving the graphql into some other file. We have a few inline query declarations that really ought to be moved out, and would cut probably 200 lines from the file.